+++ src/data/synthetic/multi_class.jl
@@ -1,0 +1,10 @@
+"""
+    load_multi_class(n=100; seed=data_seed)
+
+Loads multi-class synthetic data.
+"""
+function load_multi_class(n=100; seed=data_seed, centers=4)
+    counterfactual_data = load_blobs(n; seed=seed, centers=centers, cluster_std=0.5)
+
+    return counterfactual_data
+end
